feat(#78): add unit tests
